feat: Made ModalContainer more reusable for footer
